added order item preparation and test

+++ src/CurationService.Core/Consumers/OrderCurationConsumer.cs
@@ -19,7 +19,7 @@ namespace CurationService.Core.Consumers
 
         public async Task Consume(ConsumeContext<CurateOrders> context)
         {
-            var orderItems = _db.OrderItems.Where(x => x.Status == OrderItemStatus.Expired).ToList();
+            var orderItems = _db.OrderItems.Where(x => x.Status == (int)OrderItemStatus.Expired).ToList();
 
             if (orderItems.Any())
             {
